+++ src/main.js
@@ -1,50 +1,15 @@
-import fastjsDom from "./fastjsDom/fastjsDom";
-import fastjsDomList from "./fastjsDomList/fastjsDomList";
-import config from "./config";
-import _dev from "./dev";
+// modules
+import Ajax from "../modules/ajax/main";
 
-let fastjs = {
-    selecter(el, place = _dev._dom) {
-        let dom = []
-        // if place = fastjsDomList
-        if (place.constructor === fastjsDomList) {
-            place._list.forEach((e) => {
-                e._el.querySelectorAll(el).forEach((v) => {
-                    // check if v is in dom
-                    if (!dom.some((v2) => v2 === v))
-                        dom.push(v);
-                })
-            })
-        } else {
-            dom = place.querySelectorAll(el);
-        }
-        // if last selecter is id
-        let special = el.split(" ")[el.split(" ").length - 1].startsWith("#");
-        // prevent extra forEach
-        if (!special)
-            config.dom.specialDom.forEach((v) => {
-                if (el.startsWith(v))
-                    special = true;
-            })
-        if (special)
-            // -> fastjsDom -> element
-            return new fastjsDom(dom[0]);
-        // -> fastjsDomList -> fastjsDom -> element
-        return new fastjsDomList(dom);
-    },
-    copy(text) {
-        // copy text to clipboard
-        let input = new fastjsDom("input");
-        input.val(text).push();
-        input._el.select();
-        document.execCommand("copy");
-        console.log(input);
-        input.remove();
-    }
-}
+// src
+import FastjsArray from "../src/fastjsArray/fastjsArray";
+import FastjsDom from "../src/fastjsDom/fastjsDom";
+import FastjsDomList from "./fastjsDomList/fastjsDomList";
 
-let selecter = fastjs.selecter
-let copy = fastjs.copy
+// main
+import fastjs from "./methods";
+import {selecter, copy} from "./methods";
 
-export {selecter, copy}
-export default fastjs
+// export
+export {Ajax, FastjsArray, FastjsDom, FastjsDomList, selecter, copy};
+export default fastjs;
